pyInterface: add method fileManager.getDataFilePath()
add the method getDataFilePath to the filemanager. It returns a list of all
filenames that match the given bin information. The bin information data type
is a dictionory with the bin variable as key (e.g. 'mass') and a value between
the bin boundareis as values of the dictionary. If a boundary value is given
the method prints an error and aborts.