pyInterface: add utils function for creating a binning map from the command line argument
add a utility function that interprets the command line argument for the
binning map and converts into the standard binning map dictionary type. The
interpreter expects a list of strings, where each string has 3 semicolon
seperated elements. The first is the name of the binning variable while the
last 2 elements are the boundaries of the bin. The easiest way to get this list
is by using the argparse module and setting the action of the argument to
'append'.