pyInterface: support loading/saving of fileManager to hard drive
create functions to save/load a fileManager object to/from the hard drive.
These functions are located in pyInterface/package/_fileManager.py and are
executed by calling pyRootPwa.core.saveFileManager (or load). Create a new
script that creates a fileManager and saves it to the hard drive for later use.
The metadata eventsType enum has to be internally handled by a python class
that resembles an enum to allow pickle to save it to the hard drive (bug in
boost::python).