pyInterface: add calcIntegrals python script
add userInterface/calcIntegrals.py script. This script calculates an amplitude
integral matrix from all the amplitude files determined by the file manager.
Before executing calcIntegrals.py you need to first run calcAmplitudes.py with
the same fileManager.
For this change we had to introduce a new package function "calcIntegrals"
which takes a list of amplitude files as argument and returns the
ampIntegralMatrix object. The package function extracts all the amplitude trees
from the amplitude root files and adds the waveNames from the amplitude
metadata to the ampIntegralMatrix object (this is the new method 'addWave').
Then calcIntegrals calls the ampIntegralMatrix.integrate method, which has been
rewritten to take a vector of TTrees as argument (the tree extraction was
previously done in the ampIntegralMatrix object).