buildSystem: Fix bug in git hash getting.
Apparently, with some git installations, there was a problem when getting the
latest git hash when environment.cmake was run, because git did not recognize
the full path to the repository as belonging to the repository (*sigh*). Fix
this by running git within the repository and not specifying any path.