pyInterface: clean up config file
general clean up of the config file and the package/_config.py module.
Delete config entries "massBinDirectoryNamePattern" and
"dataAmplitudeDirectoryName". Delete amplitude category and put folders for key
and amplitude files into general section. Create category "other" and put all
weight related stuff, as well as amplitude subdirectory names in there.
Create a function that manages all the directory handling. The function checks
if the config file has the option and assigns a default if the option does not
exist. It also expands environment variables and checks if an absolute path was
given. If it recognizes a relative path it returns the path relative to the
position of the config file (and NOT relative to the current path).
Thanks to sensible defaults the config file works in theory without giving any
of the general options. The initialize method creates directories (for data, key
and amplitude files) if they do not exist yet.