NLopt: Create nlopt subfolder.
Create a new subdirectory with the copied pwafit as source file which is
enabled if NLopt is found. Add a new cmakeModule to find NLopt in CMake and
provide it for compilation.

(cherry picked from commit 61e42bb6c0bf242a51bc454ee984fad4b37856e6)

Conflicts:
	CMakeLists.txt

For the master branch, this commit has been modified so that new variable names
for the library and include directory paths is used.