NLopt: Add first functional version.
Modify pwaNloptFit.cc so that now NLopt is actually used. The algorithm which
is selected (i.e. hardcoded) is one which should be close to Migrad's behavior.
A first test seems to give sensible results. Still to be implemented are the
bounds for the algorithm and a printing of the result status of the optimizer.

(cherry picked from commit 75601d1a93688870ab3ef301863c79a94a668f72)