src: Implement half-Cauchy priors.
Implement Cauchy shaped priors for all the parameters into the pwaLikelihood.
There is a log of debug and comments still in, but there seems to be a problem
with non-converging fits which has to be fixed, then the debug code will go.

(cherry picked from commit 1e6f8190a4f57f294447a44afdccac49a8427d9b)

For the master branch, this commit has been modified so that the static const
double _cauchyWidth is declared in a c++11 compatible way.