src: Fix bugs in the Cauchy priors.
Fix several bugs in the half-Cauchy priors: the derivatives were not correctly
applied because they were mixed in with commented debug code (which made it
work whenever the printouts were switched on to debug *sigh*). Also, in FdF,
the derivatives were calculated but the likelihood was not affected by the
priors. Still a lot of debug code in, will be cleaned in the next commit.