evtweight: move creation of difference plots
Move the creation of the difference and acceptance histograms from the
plotWeightedEvts script to the plotGlobalWeightedEvts script. This way
the plotWeightedEvts is a script only filling the histograms required
later on, but not performing any other actions.

(cherry picked from commit efa5cf35197d5916de85e2b7af2617dfdef9b253)