massDepFit: extract final-state mass-dependence from model
Extract the final-state mass-dependence from the class 'model' and put
it into its own class 'fsmd'. Similar as for the 'component' class, the
new 'fsmd' class takes care to 'init' and 'update' itself from and to
the configuration file. Also the parameters are now handled in a very
similar way. The TF1 previously used has been replaced by a TFormula as
it wil never be drawn. The evaluation of the formula is now done using
the 'EvalPar' method (no longer 'Eval') thus it is not necessary to
call 'SetParameter' for the TF1/TFormula. Fixed variables are now taken
care of by the minimizer.