massDepFit: Fix comparison between signed and unsigned numbers
GCC 4.4 throws an error when comparing signed and unsigned numbers, so
do not do this. At the same time this fixes the case that an object of
the class 'fsmd' is used despite the 'init' method returning false. In
this case the number of parameters stored in '_nrParameters' might not
be the correct number of entries in the '_parameters*' vectors.