massDepFit: first branching is not a fit parameter
Do not create a fit parameter for the first branching. This branching is
in any case real and fixed to 1. This is ensured while reading the value
from the configuration file, there should be no (legal) possibility to
modify this value.
(cherry picked from commit 2910177e2f4fdf547246c6bf193b2726c874cbe6)