src: Strip covariance matrix too.
Add an option to stripIntegralsFromFitResult which adds the possibility to
delete the covariance matrix as well. This is useful to save disk space in
cases where the covariance matrix is a dummy matrix anyway (i.e. when the file
was written with older versions of pwaNloptFit).

(cherry picked from commit 3b9fcd66c853f54e669eee7f8080c002ee2b366c)