massDepFit: add '_id' member to components
Add a new member '_id' to the 'component' class. This ID should be a
unique index for this component, for the moment the position in the
components vector of the fit model is used (unless this is sorted
anywhere (which it is not)).
(cherry picked from commit d83c173c81e7987c4322d7bc68e476019c21b7db)