massDepFit: remove fit parameters from 'component' class
Remove the fit parameters from the 'component' and 'channel' classes. For
this a new 'parameters' class was added that simply keeps arrays of the
various non-constant information required during the fit. This new object
is filled by the 'likelihood' and 'model' classes, all functions that
require input from fit parameters now require a reference to such an
object as an additional parameter.

This should allow to evaluate the chi2 value from multiple threads
without interference, as all the changing stuff is contained in
'parameters'.