massDepFit: integral for Bowler
No longer get the phase-space integral used in the Bowler
parameterisation from the corresponding channel, but require it to be
set in the configuration file with a new option 'integral' (a list of
arrays of two doubles). This should be a bit more flexible when it comes
to, e.g., testing the various possibilities how to correctly take the
overall phase-space into account.