massDepFit: fix upper limit of for loop
Fix a potential endless loop while updating the configuration file, the
for loop was not terminated if the loop variable exceeded an upper
limit. As the configuration file was correctly read before, this can
only occur if the in-memory content of the configuration file is
modified between reading and writing, i.e. you messed up memory
managment pretty badly.