resonanceFit: rename Bowler parameterization
Rename the component type 'parameterizationA1Bowler' to
'integralWidthBreitWigner' which is based on the same principle. The
phase-space integral is used to determine a mass-dependent width.

The new name admittedly is a stupid name, but I hope the analogy to the
'dynamicWidthBreitWigner', where the dynamic factors are the break-up
momentum and the barrier factors, excuses this choice.
(cherry picked from commit d75ae274fd90a67f4e4ac8031835721c1b1d7f95)