resonanceFit: move per channel options in parameterizations
Move the per channel options like the masses of the isobars, the
relative orbital angular momentum and the branching ratio from the
component to the individual decay channel section for the
'dynamicWidthBreitWigner' and the 'integralWidthBreitWigner'. This
should make it more simple to keep track of the values used for those
quantities.