resonanceFit: move per channel options in parameterizations
Move the per channel options like the masses of the isobars, the
relative orbital angular momentum and the branching ratio from the
component to the individual decay channel section for the
'dynamicWidthBreitWigner' and the 'integralWidthBreitWigner'. This
should make it more simple to keep track of the values used for those
quantities.
(cherry picked from commit 5d0c5819b3e5f434531bc9da3db38eb15dba5666)