resonanceFit: skip dynamic factors if branching is 0
For the 'dynamicWidthBreitWigner' skip the calculation of the dynamic
factors if the branching ratio is zero. The zero weight in the sum will
in any case lead to this summand being ignored.

In the 'integralWidthBreitWigner' the interpolation is also not
performed in case the branching ratio is zero.
(cherry picked from commit 0659128d212d3905b3e4d82a463b793c103480a0)