resonanceFit: fit to prod. amp. using only errors on real and imag part
Allow the fit to the production amplitudes to be performed using only
the errors on real and imaginary part from the mass-independent fit
(ignoring the rest of the covariance matrix). This required to replace
the previously used boolean variable with an enum that can take several
values to take care of this.

The constructor of 'massDepFitLikeli' sets a default behaviour if
nothing has been provided at the command line. The 'init' method checks
whether all settings are compatible.