src: Change anchor/flat wave behaviour in Likelihood
After finishing the minimization change the signs of the anchor- and flat wave
so that they are forced to have a positive amplitude. Change the signs of the
other amplitudes accordingly. Also remove the start value constraints for the
anchor wave to be positive since the way of forcing the amplitude to be
positive was inneffective (large step sizes in the early minimization
iterations made the amplitude negative regardless of start value). This has to
be done for both fit executables.