resonanceFit: enable thread_local cache
When ROOTPWA is compiled with the C++11 standard enabled, a thread
local cache is used. This is basically a static variable for each
thread, so that each thread can continue to use the production
amplitudes (and later possibly other quantities) calculated before. The
thread local storage ensures that this also works if the likelihood
function is called in a multithread environment.
(cherry picked from commit 666758daa83578d439127cae16aeeed502187272)