NLopt: Add command line options for saving space.
To make fits with a lot of waves (>1000) and a lot of attempts (>100) feasible,
introduce an option to save disk space when fitting. If this option is set in
pwaNLoptFit, neither the covariance nor the integral matrices are saved to
disk, drastically reducing the disk space needed.

(cherry picked from commit 10407d978dee5f98933fdddfffee532033795e8f)