resonanceFit: bugfix for fit using branchings
While performing a fit using branchings, the parameters were not set
correctly. The number of parameters and the ordering of the parameters
might have been wrong in this case. Now only loop over the number of
couplings and branchings required for the fit and do no longer try to
find out for each channel whether it should be used to define a
parameter or not.
(cherry picked from commit dabd1018730e9a5861ca39e969ab0fa2f73dd411)