resonanceFit: bugfix for reading config file when using branchings
Fix a bug that occured while reading the configuration file when
branchings were used. It is perfectly valid that branchings are not
read, and the channel does not use the same branching as the very first
channel of this component.

This requires to in the end check that indeed at least one branching was
read (otherwise one needs to be set in the 'parameters' object).

More explicitley use branching/coupling index instead of the respective
counter (which should be the same).