NLopt: Abort nlopt fit if thresholds are encountered.
As the nlopt package cannot handle fixed values and problems arise with the
calculation of the covariance matrix, prevent fitting with thresholds in the
wave list by exiting with an error message.

This should be fixed in the version 3 of rootpwa but will not be backported to
the _v2 branch.

(cherry picked from commit b57fe56c3c84ebc3a7a56634cd6104f8d7030d06)