src: Fix error printing in pwafit.
Fix the errors which are printed after the minimization in pwafit. While the
covariance matrix is multiplied with 0.5 to accommodate for the fact that we
are fitting a log(Likelihood), the errors which are directly taken from the
minimizer lack this correction factor. Add the factor sqrt(0.5) (because the
errors are the square root of the diagonal of the covariance matrix) to the
errors being printed by pwafit.

(cherry picked from commit effba91b68291ce904a0232ab629afe37bafbfc0)