NLopt: Fix parameter printing and print errors.
Fix the formatting of the parameter printing in pwaNloptFit and also print the
errors for the fit parameters. Print the errors even if the covariance matrix
is not saved to the fit result file so that they are at least available in the
printout in this case.

(cherry picked from commit 0cd51f4c00dc5833eb6157a313476831b7c22b28)