NLopt: Do not calculate the Hessian in case of "-z".
If the "-z" option is set in pwaNloptFit (i.e. the covariance and integral
matrices are not saved to the fit result file), do not calculate the Hessian
matrix. This saves quite some computing time if the fit has a lot of parameters
(which is to be assumed if the matrices are not written to disk) but has the
downside that the eigenvalues of the Hessian cannot be checked.