Src: Fix bug in fitResult's prodAmpCov().
Check if the covariance matrix is valid in fitResult::prodAmpCov() before
trying to access it. If the covariance matrix is not valid, return the 0-matrix
with the correct dimension, so scripts expecting to get a 2x2 matrix do not
crash (at least not because of matrices of the wrong size).