resonanceFit: fix chi2 for fit to production amplitudes
Fix the definition of the chi2 for the fit to the production
amplitudes. The value might have been wrong, if different mass limits
were applied to different waves. In this case the covariance matrix of
all non-zero waves was used, while only the covariance matrix
corresponding to the waves used in the current bin should have been
used.

This is now fixed, special care had to be taken in case the anchor wave
was excluded due to this mass limits.

In principle also the fit to the spin-density matrix could have been
affected, but due to the block structure of the covariance matrix (only
blocks of 1x1 or 2x2 elements), the inverse matrix would never be
affected by superflous waves, just some lines and columns would
actually contain elements not equal to zero, which would in any case be
ignored as the corresponding entries in the difference vector were 0.
Also this method using the matrices is not used here at the moment, as
it is too slow.
(cherry picked from commit fea5a7bc5c02c765a099619cbdbdcf0751aaf15c)