NLopt: Set hasHessian in fit result for nlopt fits.
Set the hasHessian attribute of the fit result written by pwaNloptFit to true
in case the fit converged and the analytically calculated covariance matrix has
been added to the fit.

Note that this does not exactly match the behavior of pwafit, which might have
hasHessian set to true even if the fit did not converge. In any case, the
number and meaning of the status booleans saved to the fit result has to be
revisited, but solving this probably involves an interface change.

(cherry picked from commit 61da28a72d730d2583b1b120b26d08779ba9628d)