src: Change behavior of "converged" in pwafit.
Set converged to false in pwafit if the calculation of the Hessian matrix by
Minuit fails.

The logic behind this change is that the converged flag now gives information
about the usability of the fit result. If anything fishy or strange was going
on either in the fit or in the calculation of the errors, converged will be
false, so that such fit results can be rejected. The hasHessian flag on the
other hand will just hold the information about the presence of a valid Matrix
and may be false even though there is nothing wrong with the fit (e.g. the
calculation of the Hessian was not performed to save time or disk space). So
hasHessian should be consulted if one wants to access errors or the evidence.
This scheme should make the meaning of these two variables more transparent and
facilitate filtering of fit results.