decayAmplitude: fix segfault when ROOT waves are skipped
Fix a segmentation violation that happened if a wave from a ROOT file
was skipped because
* there already was a wave with the same name in the integral
* the number of events differed wrt. to the previous waves
The vector 'ampTrees' did not respect this and contained NULL pointers
at the end (initially meant for those skipped waves). Now the vector is
resized to remove those entries, according to the C++ standard no
reallocation will happen, so all pointers remain valid.