src: Further improve plotGui.
Add more improvements to the plotGui: now one can have an option to only have
the waves selected which have an intensity above a given threshold. Also, there
is a possibility to give a list of waves which should be presented to be
selected. These features are barely tested and the implementation is very
clunky at the moment, but commit anyway to have something to fall back on.
(cherry picked from commit f7aa00874502ade737903bb61181f101ac7c660b)