src: Fix the wave selection in the plotGui.
Fix the wave selection with thresholds and a white-list in the plotGui so that
now a wave is either to be presented in the selection box if it is above a
given thresholds or if it appears in the white-list. Also make the whole thing
a bit more elegant.
(cherry picked from commit 57c98f7f9f301cc1bcdc25403344a4da3c93629b)