src: Fix the canvas logic in plotGui.
Remove unneeded pieces of code from plotGui and add some logic to the drawing
of canvases. Now, one can use a selection box to decide if a new plot should be
drawn in a entirely new canvas or if the existing canvas should be updated.
Thereby, add some code to make sure one does not try to update a closed canvas.