src: Fix the evidence calculation for some cases.
Fix the evidence calculation for the case of large covariance matrices (i.e.
several hundred parameters). Because of numerical limitations, the determinant
of the covariance matrix could become infinite, ruining the evidence
calculation. Fix this by calculating the eigenvalues of the covariance matrix
and adding up the logarithms of them which could still yield a sensible number.