src: Add likelihoodPointCalculator.
Add the likelihoodPointCalculator which allows to get the likelihood value for
a given set of production amplitudes from a fit result in the likelihood
function given by a wavelist. If there are more waves in the wavelist than in
the fit result, the corresponding amplitudes are set to 0.

(cherry picked from commit 22334cea86fd4627daac5dec077fa9ada6019b38)

Conflicts:
	src/CMakeLists.txt