src: Fix a bug in pwaLikelihood's readDecayAmplitudes.
Fix a bug in pwaLikelihood::readDecayAmplitude, where the array containing the
decay amplitudes was resized every time a new amplitude file was read in, while
slowed down the initial loading of amplitudes by some order of magnitudes.
While doing that, also finally fix the bug that the firstWave boolean was never
set to false so the consistency in terms of number of events in the amplitude
files was never properly checked as was intended.