src: reduce number of variable with same content
Reduce the scope of the 'nmbEvents' variable to the end of the loop
reading the amplitude files. The variable '_nmbEvents' is used to define
the size of the vector to keep the amplitudes for every wave but the
first one.

While copying the amplitudes to the multi_array used during fitting,
only copy the first 'nmbEvents' events in case '_nmbEvents' would be
larger.