src: Use wave list from fit result.
In the likelihoodPointCalculator, use the wave list for the fit result instead
of requiring a wavelist to be given on the command line. The way to do this is
quite braindead, as the likelihood requires the wavelist to be given as a file
name it then opens, so in order to do this change, a temporary file has to be
created where the waves are written to, then the likelihood reads this file and
afterwards, it is deleted again.