src: Simplify getting intensities.
Simplify getting intensities and their errors in the case where only a single
wave is being read. This is necessary as it might be possible that the integral
matrices are not loaded from file or were not present to begin with, leading to
a crash with the previous implementation despite the fact that the intensity
for a single wave can be returned without needing access to the integral
matrices.