partialWaveFit: fix time measurement in 'pwaLikelihood'
The time supposed to be spent evaluating the likelihood part of the
function in 'DoDerivative' and 'Gradient' (with USE_FDF) was summing
up the sum of the total time spent in 'DoDerivative' or 'Gradient' so
far. As 'DoDerivative' or 'Gradient' is either looking up a value in
the cache, or calling 'FdF', the values of 'normTime' (not used
anyways) and 'funcTime' are irrelevant, so remove it and only keep
track of the total time.
(cherry picked from commit 785d7f8f1c100930c594584aa04867b13f664fa3)