partialWaveFit: remove ordered parameter methods
To avoid confusion in the future get rid of using the ordered parameter
indices and the accordingly ordered wave names. They provide little
usage (only printing the parameters in the order of the waves appearing
in the wavelist), and are not used.
(cherry picked from commit 22c43955edc23e3ca752595ef892b918798ea61d)