partialWaveFit: reflectivities in integral matrices
Do not set elements of the integral matrices corresponding to pairs of
waves with different reflectivities while importing the values from an
'ampIntegralMatrix' object, the affected elements then remain zero.

This somewhat mitigates the impact of issue #25 as now at least the
overlap between waves of two different reflectivities is zero, so
intensities are calculated correctly.