partialWaveFit: protect evidence calculation for missing matrices
Protect the calculation of the evidence components against empty
covariance and accepted normalisation matrices. In case any of these
matrices is of zero size (old fit result, matrices stripped, ...),
return components that sum up to negative infinity. Make sure the number
of evidence components is the same as for a valid result.