partialWaveFit: evidence with multi-ranks and thresholds
Fix the evidence w.r.t. fits with more than one rank, more than one
reflectivity and thresholds. Instead of simply taking the determinant of
the complete acceptance integral matrix, create a sub-matrix for each
rank and reflectivity with the waves actually in the fit and calculate
the determinant thereof. This fixes issue #31.