src: Change plots of plotGui.
Change the size and layout of the canvas showing the results in the plotGui. Do
not show the real and imaginary parts anymore, as they are seldom useful and
move the phase so that it is below the first referenced intensity. This way,
features in the intensity can be easier connected to features in the phase.