partialWaveFit: helper function to extract Eigenvalues
Add a helper function to extract Eigenvalues and -vectors from the
covariance or Hessian matrix taking into account fixed parameters. The
fixed parameters are removed from the matrices before the calculation
of the Eigenvalues and -vectors, zeros are inserted into the
Eigenvectors corresponding to the fixed parameters.

Use this new function, 'eigenvectorLikelihoodSlices' now can make use
of fits with fixed parameters.