buildSystem: test for thread_local support
'Apple LLVM 6.1.0' does not support 'thread_local' storage even with
C++11. The '-std=c++11' compiler switch cannot simply be removed, as
it is also required for ROOTv6. So the preprocesser cannot check
whether C++11 is activated to decide whether to use 'thread_local' or
not.

Now test for 'thread_local' while running cmake, and define
'COMPILER_PROVIDES_THREAD_LOCAL' if it does.