buildSystem: find BAT via bat-config
If the environment variable BATINSTALLDIR is not set, try to find BAT
via 'bat-config'. Try to extract a valid path from
'bat-config --prefix'. The output of 'bat-config --version' (for the
version string) and 'bat-config --libs' (to determine whether to add
'-fopenmp' to the linker flags) is preferred over parsing the
(internal) 'config.h' header file.