src: Make width of half-Cauchy prior settable.
Make the width of the half-Cauchy prior changeable at runtime by making it a
standard member variable of the pwaLikelihood (as opposed to a const static
one) and adding a getter and setter for it. Also introduce a new command line
option in pwaNloptFit.cc which allows to set the prior width.
(cherry picked from commit 3ccfc38b5af58eb507795e59b38dfc335da27ca4)