resonanceFit: fill 'fitParameter' object with errors
While reading the initial values and errors for the parameters from
the configuration file, fill the errors into a 'fitParameter' object.

Also chi2 and ndf are now read from the file. For storing chi2 and ndf
adopt a scheme similar to the reading, namely move the filling ob the
'fitquality' group into an own method.