resoanceFit: shuffle reading of configuration file
Shuffle the parts for reading the configuration file, the goal was
that each function can require not being called with a NULL pointer,
the cases where this could happen on purpose are now handled one level
before (in the calling method). Also make sure that further methods
are only called if indeed the required option was found.