pyInterface: support wavelists in calcAmplitudes script
add a wavelist command line argument. Only calculate the amplitudes of the
waves that appear in the wavelist. If no wavelist is given calculate all
amplitudes just like before. Still have to implement thresholds on this level,
they are ignored right now and amplitudes are calculated even if they are sub
threshold.