resonanceFit: handle parameters in FSMD similar to components
Handle the parameters in the final-state mass dependence similarly to
the way they are handled for components. Do not read the values from
an array, but have a group for each parameter. Only the 'val' and
'fix' variables are required, the other variables 'upper', 'lower',
'step', and 'error' (this particular one without any effect) are read
whenever available.

Update the example configuration file accordingly.