highLevelInterface: add interface to partial-wave fit
Provide an interface to the partial-wave fit in the high-level
interface, and also provide Python bindings for this. This is later to
be used from Python scripts, for the moment the 'pwafit' executable
has been modified accordingly. Due to changes to the 'pwaLikelihood'
class, some other executables are no longer compiled.

Due to limitations of the number of arguments on BOOST's Python
wrapper not all parameters can be set for the NLopt fit.