pyInterface: create pwaFit python executable
create a python executable that has similar functionality to the pwafit.cc
C++ executable. Do all the file handling in a package function (fit.py) and
call a highLevelInterface C++ function (pwaFit.cc) with C++ objects as
arguments. Since some of the file handling was done in pwaLikelihood.cc some
changes are needed in the Likelihood as well. Consequently almost all
executables are unusable at the moment (hence all the changes in CMake files).
The executables will have to be recreated in python and moved to appropriate
directories (legacyScripts) in the following commits.