relampl: Move to vectors in TFracNum.
Change the data type for holding the exponents of the prime numbers from arrays
to vectors in TFracNum. This required a rewrite of much of the TFracNum class,
but seems to give the same results. Several small things still have to be
ironed out in TFracNum, but I think there is now a reasonable basis.