relampl: Add constness in TFhh.
Make pointers in arguments const. It is not entirely clear why this is working
and if this constness can be broken: after all, a vector of pointers is copied
out of the const objects and could potentially be changed by the callee. At
some point, one should probably change over to smart_ptrs, but of course they
have their problems with const-ness as well...