Fix warnings coming from -Wextra.
Fix several warnings which popped up when compiling with -Wextra. Some of them
required making an educated guess as to the correct solution, it remains to be
seen if the way chosen turns out to be well suited.

(cherry picked from commit c96551881decae4dda87cba78a0dba2ed1b4986d)

In addition to fixing the conflicts below, also fix all the other compilation
errors in the master branch with -Wextra added.

Conflicts:
	resonanceFit/massDepFit.cc
	resonanceFit/massDepFitComponents.h
	resonanceFit/massDepFitFsmd.cc
	tools/addisokey.cc
	tools/key2tex.cc