buildSystem: restrict ignoring of warnings for ROOT dictionaries
Restrict the ignoring of warnings while compiling the ROOT
dictionaries a bit. So far it has only been observed with one version
of Apple's clang, so for the moment only ignore this warning for this
compiler. From the ROOT repository it also seems that this will be
fixed in the next release (5.34.32), so only ignore the warning for
earlier versions.