resonanceFit: parameters for final-state mass-dependence by name
Use the name of the parameters as used in ROOT to access and write the
parameters from and to the YAML file. For ROOT5 this is not supposed
to change anything, for ROOT6.04 this allows to define formulas like
        "[massThr]+[abc]*x"
and have parameters "massThr" and "abc" in the config file (instead of
having to use numbers [0], [1], ... and "p0", "p1", ...). This also
fixes an issue in ROOT6.04 with formulas where the parameters have
holes, i.e.
        "[0]+[2]*x"