pyInterface: Add python interface for pwaLikelihood class
Add a python interface for the pwaLikelihood class to be able to initialize the
likelihood in a python script. Change the userInterface executables (pwaFit and
pwaNloptFit) to take a likelihood object as an argument and change the
highLevelInterface C++ functions (again pwaFit and pwaNloptFit) to take a
likelihood object as well. Put filehandling back into the likelihood
initialization since opening the files in python uses too much memory.