decayAmplitude: fix phase-space integral
In the estimation of the phase-space integral the integration volume
was taken into account twice, once in the normalization of the
'S_U_CHUNG' weight, and once in the calculation in
'phaseSpaceIntegral.cc', so the later one was removed. Also
explicitely set the current default weight and kinematic types to
ensure they always watch with the requirements in
'phaseSpaceIntegral'.