decayAmplitude: fix numerical instabilities in gjTransform
Fix numerical instabilities in 'isobarAmplitude::gjTransform' when the
beam and X Lorentz-vectors are equal. In this case the production
plane is not well defined, and any rotation might skew up the angles
as one particular plane might be prefered due to numerics. Boosting is
in principle possible, but will print a warning if it is not
negligible.