pwaLikelihood: protect against zero entries in norm. integral
Normalization integral entries that are zero will lead to NaNs
appearing in (normalized) amplitudes and subsequently the likelihood.
This might happen in case of, e.g., the 'flatRange' mass dependence,
however in this case also the amplitude should be zero. So in case the
normalizetion integral entry is zero, also all according amplitude
values have to be zero, otherwise an error is triggered.