pyInterface: remove includes from header files
Where possible removed include statements from header files and added
them to the corresponding source file. For basically all classes it is
not necessary to have the full C++ class available if only the
'export...' function is to exported and to be called, this might
somewhat reduce the used memory.