decayAmplitude: add and implement mass dependence for freed-isobars
Add rectangular mass dependence for freed-isobar fits. The
'waveDescription::mapMassDependenceType' method now takes a libconfig
object as argument as the new mass dependence cannot be fully
initialised from just the name.

As for this mass dependence several waves differ only in the 'bounds'
argument the wave name created with the previous method would have
returned the same name for multiple waves. So for the new mass
dependence not the (particle) name of the isobar is written, but a
combination of the quantum numbers and the bounds is created.