partialWaveFit: Overload fitResults's fill method.
Overload the fill() method of the fitResult class so that it can also take
another fit result as argument. This is useful if one intends to save multiple
fit results to a tree in python without (re)setting the branch address every
time or deal with python's way of copying objects by references.