decayAmplitude: consistent mass-dependence names
Make the mass-dependence names consistent between the mapping done in
'waveDescription::mapMassDependenceType' and the names written to the
keyfiles which are defined in the 'name' method of the massDependence
classes. For 'relativisticBreitWigner' and 'flatMassDependence'
changed the name in the mapping because those names are typically not
used as they are default values. Also the names consistent with the
mass-dependence are already used in one other place. For the other
mass-dependencies change the name in the class as there the names are
already used.