decayAmplitude: handling of mass-dependencies when writing keyfiles
The mass dependence names 'flatMassDependence' and
'relativisticBreitWigner' are special as they are the default for the
X, respectively any isobar, decay. If a mass dependence was of one of
those two types, the 'massDep' group was never added to a keyfile,
however, it might be possible (unlikely as it seems) that an isobar
decay should indeed use the 'flatMassDependence'. So now also the
type of the decay is checked before assuming the name to be the
default mass dependence.

Also fix the default value for the boolean determining whether to
write the production vertex to the keyfile in the Python interface,
it now is the same as for the C++ equivalent.