decayAmplitude: change way key files are parsed
Change the way key files are processed by 'waveDescription'. The
methods 'parseKeyFile*' are made static, and are now returning the
'waveDescription' object.

The Python interface has been adjusted accordingly, the test for the
constructor is no longer useful (as the constructor itself is now
basically useless (but required for ROOT I/O)).