partialWaveFit: Clean up wave name escaping.
Clean up the business with the pattern matching for wave names in the fit
result by removing some ancient lines, making the escaping functions available
in python and adjusting plotGui accordingly. This probably breaks some of the
root scripts using the fit result, but all the compiled code seems to continue
to work.