storageFormats: print amplitude metadata from Python script
Modify the Python version of 'printMetadata' to be able to also print
amplitude metadata and not only event metadata. A new option '-a' has
been added allowing to the the base name of the amplitude object
(typically the wave name). If this option is not present, try to
guess the name from filenames following the naming convention of the
file manager.

Also removed the initialisation of the printing stuff which since
some time is taken care of during 'import pyRootPwa'.