pyInterface: calculate covariance matrix from fit result
Add a python script that takes a fit result and calculates the
covariance matrix for this result. It is similar to the C++ executable
with the same name that stopped working when switching to the Python
user interface.

The script relies on the file manager to provide the key, amplitude and
integral files. The correct ID of the bin to calculate for has to be
provided. The wavelist is extracted from the fit result.