partialWaveFit: move eigenvalues and -vectors to likelihood
Move the functions to calculate the Eigenvectors and Eigenvalues of
the Hessian from the 'partialWaveFitHelper' stuff into the
'pwaLikelihood' class. At the moment the function calculating the
Eigenvectors and -values requires the Hessian to be present as an
argument, but this can simply be extended to work from raw (fit)
parameters. Also add the corresponding function in the Python
interface.