pyInterface: 'get(Best)FitResult(s)FromFile' globally available
Remove the 'getFitResultFromFile' function from the file
'calcCovMatrixForFitResult.py' and instead implement it in the
'_fitTreeUtils.py' so it can be used elsewhere as well, export this
function in the 'pyRootPwa.utils' namespace. This function takes the
file name of a ROOT TFile containing a TTree with a single entry of a
converged fit result. If those conditions are not met, 'None' is
returned.

Remove the 'getBestFitResult' function from 'genPseudoData.py' and
also move it to '_fitTreeUtils.py', export it, and rename it to
'getBestFitResultFromFile'. This function takes as arguments a
filename and (optionally) a mass. If the mass is specified, find the
fit result that has the smallest mass difference to the given mass,
and if there are several also select the best likelihood. Only
converged fit results are considered. If the mass is not given,
simply search for the best fit result, if several mass bins are in
the file, emit a warning.

Also add a 'getBestFitResultsFromFile' function that finds the best
converged fit in each mass bin.

'getBestFitResultFromFile' does not make use of
'getBestFitResultsFromFile', as a copy of unneeded fit results would
need to be kept in memory, which might be a problem if huge fit
results are present in a lot of bins. The previous functions have
been remove for the same reason, they relied on being able to keep
all fit results in a file in memory.